Added interfaces for all components like @Service, @RestController, @Component
