--- address-etl/src/test/java/com/rmamedov/deasy/addressetl/AddressEtlApplicationTests.java
@@ -1,13 +1,0 @@
-package com.rmamedov.deasy.addressetl;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AddressEtlApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
