+++ address-etl/src/test/java/com/rmamedov/deasy/addressetl/ContextLoadsTest.groovy
@@ -1,0 +1,19 @@
+//package com.rmamedov.deasy.addressetl
+//
+//import org.springframework.beans.factory.annotation.Autowired
+//import org.springframework.boot.test.context.SpringBootTest
+//import org.springframework.context.ApplicationContext
+//import spock.lang.Specification
+//
+//@SpringBootTest(classes = AddressEtlApplication.class)
+//class ContextLoadsTest extends Specification {
+//
+//    @Autowired
+//    private ApplicationContext context
+//
+//    def "When application starts then context's loaded"() {
+//        expect:
+//        context
+//    }
+//
+//}
