+++ order-service/src/main/java/com/rmamedov/deasy/orderservice/service/OrderService.java
@@ -1,49 +1,17 @@
 package com.rmamedov.deasy.orderservice.service;
 
-import com.rmamedov.deasy.model.exceptions.OrderNotFoundException;
 import com.rmamedov.deasy.orderservice.model.repository.Order;
-import com.rmamedov.deasy.orderservice.repository.OrderRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-import java.util.Set;
+public interface OrderService {
 
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class OrderService {
+    Mono<Order> save(Order order);
 
-    private final OrderRepository orderRepository;
+    Mono<Order> findByCriteria(String id, String checkStatus, String paymentStatus);
 
-    @Transactional
-    public Mono<Order> save(final Order order) {
-        return orderRepository.save(order);
-    }
+    Mono<Order> findById(String id);
 
-    @Transactional(readOnly = true)
-    public Mono<Order> findByCriteria(final String id,
-                                      final String checkStatus,
-                                      final String paymentStatus) {
-
-        return orderRepository.findByIdAndCheckStatusesInAndPaymentStatus(id, Set.of(checkStatus), paymentStatus)
-                .switchIfEmpty(Mono.error(new OrderNotFoundException(
-                        String.format("Order with: id '%s', checkStatus '%s', paymentStatus '%s' - Not Found", id, checkStatus, paymentStatus)
-                )));
-    }
-
-    @Transactional(readOnly = true)
-    public Mono<Order> findById(final String id) {
-        return orderRepository.findById(id)
-                .switchIfEmpty(Mono.error(new OrderNotFoundException(String.format("Order with id: '%s' - Not Found", id))));
-    }
-
-    @Transactional(readOnly = true)
-    public Flux<Order> findAll() {
-        return orderRepository.findAll();
-    }
+    Flux<Order> findAll();
 
 }
