+++ order-service/src/test/groovy/com/rmamedov/orderservice/service/CheckOrderServiceImplImplTest.groovy
@@ -14,8 +14,8 @@ import com.rmamedov.deasy.orderservice.converter.OrderToOrderStatusInfoConverter
 import com.rmamedov.deasy.orderservice.converter.OrderToOrderStatusInfoConverterImpl
 import com.rmamedov.deasy.orderservice.model.controller.OrderCreateResponse
 import com.rmamedov.deasy.orderservice.model.repository.Order
-import com.rmamedov.deasy.orderservice.service.CheckOrderService
-import com.rmamedov.deasy.orderservice.service.OrderService
+import com.rmamedov.deasy.orderservice.service.CheckOrderServiceImpl
+import com.rmamedov.deasy.orderservice.service.OrderServiceImpl
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.boot.test.mock.mockito.MockBean
@@ -30,7 +30,7 @@ import static org.mockito.ArgumentMatchers.any
 import static org.mockito.Mockito.when
 
 @ContextConfiguration(classes = [
-        CheckOrderService.class,
+        CheckOrderServiceImpl.class,
         KafkaSenderConfig.class,
         KafkaSenderConfigurationProperties.class,
         TopicConfigurationPropertiesConfig.class,
@@ -38,13 +38,13 @@ import static org.mockito.Mockito.when
         OrderToOrderStatusInfoConverterImpl.class,
         OrderCreateRequestToOrderConverterImpl.class
 ])
-class CheckOrderServiceTest extends Specification {
+class CheckOrderServiceImplImplTest extends Specification {
 
     @MockBean
-    private OrderService orderService
+    private OrderServiceImpl orderService
 
     @Autowired
-    private CheckOrderService checkOrderService
+    private CheckOrderServiceImpl checkOrderService
 
     @Autowired
     private OrderCreateRequestToOrderConverter requestToOrderConverter
